# docs/swagger.yaml
basePath: /v1
definitions:
  controller.pushRequest:
    properties:
      body:
        type: string
      deviceToken:
        type: string
      title:
        type: string
    required:
    - body
    - deviceToken
    - title
    type: object
  dto.AnnouncementResponse:
    properties:
      content:
        type: string
      createdAt:
        type: string
      isBanner:
        type: boolean
      title:
        type: string
    type: object
  dto.AppVersionResponse:
    properties:
      createdAt:
        type: string
      downloadUrl:
        type: string
      isForceUpdate:
        type: boolean
      platform:
        type: string
      releaseNotes:
        type: string
      versionCode:
        type: integer
      versionName:
        type: string
    type: object
  dto.CampusMapFeatureResponse:
    properties:
      geometry:
        $ref: '#/definitions/model.FeatureGeometry'
      properties:
        $ref: '#/definitions/model.FeatureProperties'
      type:
        type: string
    type: object
  dto.CampusMapResponse:
    properties:
      features:
        items:
          $ref: '#/definitions/dto.CampusMapFeatureResponse'
        type: array
      type:
        type: string
    type: object
  dto.CheckAppVersionResponse:
    properties:
      hasUpdate:
        type: boolean
      isForceUpdate:
        type: boolean
      latestVersion:
        $ref: '#/definitions/dto.AppVersionResponse'
    type: object
  dto.ElectricityTaskOption:
    properties:
      building:
        type: string
      campus:
        type: string
      notifyTime:
        type: string
      room:
        type: string
    required:
    - building
    - campus
    - notifyTime
    - room
    type: object
  dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  dto.LoginRequest:
    properties:
      token:
        type: string
    required:
    - token
    type: object
  dto.LoginResponse:
    properties:
      profile:
        $ref: '#/definitions/sso.Profile'
      token:
        type: string
    type: object
  dto.SemesterCalendarDetailResponse:
    properties:
      calendarEnd:
        type: string
      calendarStart:
        type: string
      customWeekRanges:
        items:
          $ref: '#/definitions/model.CustomWeekRange'
        type: array
      notes:
        items:
          $ref: '#/definitions/model.CalendarNote'
        type: array
      semesterCode:
        type: string
      semesterEnd:
        type: string
      semesterStart:
        type: string
      subtitle:
        type: string
      title:
        type: string
    type: object
  dto.SemesterCalendarListResponse:
    properties:
      semesterCode:
        type: string
      subtitle:
        type: string
      title:
        type: string
    type: object
  dto.SyncElectricityTaskRequest:
    properties:
      deviceToken:
        type: string
      tasks:
        items:
          $ref: '#/definitions/dto.ElectricityTaskOption'
        type: array
    required:
    - deviceToken
    type: object
  model.CalendarNote:
    properties:
      content:
        type: string
      needNumber:
        type: boolean
      row:
        type: integer
    type: object
  model.CustomWeekRange:
    properties:
      content:
        type: string
      endRow:
        type: integer
      startRow:
        type: integer
    type: object
  model.FeatureGeometry:
    properties:
      coordinates:
        items:
          items:
            items:
              format: float64
              type: number
            type: array
          type: array
        type: array
      type:
        type: string
    type: object
  model.FeatureProperties:
    properties:
      campus:
        type: string
      category:
        type: string
      name:
        type: string
    type: object
  sso.Profile:
    properties:
      avatar:
        description: 最终头像链接
        type: string
      categoryName:
        description: 学生类别
        type: string
      certCode:
        description: 身份证号（打码）
        type: string
      defaultUserAvatar:
        description: 默认头像链接
        type: string
      deptName:
        description: 学院名称
        type: string
      email:
        description: 邮箱
        type: string
      headImageIcon:
        description: 用户设置的头像链接
        type: string
      phone:
        description: 手机号（打码）
        type: string
      userAccount:
        description: 学号
        type: string
      userName:
        description: 姓名
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: go-csust-planet 项目的 API 接口文档
  title: go-csust-planet API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: 使用Token进行登录，如果用户不存在则自动注册。登录成功后返回含有JWT的凭证信息。
      parameters:
      - description: 登录请求，需包含获取的token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 用户登录
      tags:
      - auth
  /config/announcements:
    get:
      description: 获取当前生效的公告列表，按创建时间倒序排列
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.AnnouncementResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 获取公告列表
      tags:
      - config
  /config/app-versions:
    get:
      description: 获取指定平台的所有App版本历史
      parameters:
      - description: 平台(ios或android)
        enum:
        - ios
        - android
        in: query
        name: platform
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.AppVersionResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 获取App所有版本
      tags:
      - config
  /config/app-versions/check:
    get:
      description: 检查指定平台的App是否有更新
      parameters:
      - description: 平台(ios或android)
        enum:
        - ios
        - android
        in: query
        name: platform
        required: true
        type: string
      - description: 当前版本号
        in: query
        name: currentVersionCode
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CheckAppVersionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 检查App版本更新
      tags:
      - config
  /config/campus-map:
    get:
      description: 获取GeoJSON格式的校园地图数据
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CampusMapResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 获取校园地图数据
      tags:
      - config
  /config/semester-calendars:
    get:
      description: 获取所有校历的列表，按学期代码倒序排列
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.SemesterCalendarListResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 获取校历列表
      tags:
      - config
  /config/semester-calendars/{semester_code}:
    get:
      description: 根据学期代码获取该学期的详细校历信息
      parameters:
      - description: '学期代码(如: 2024-2025-1)'
        in: path
        name: semester_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SemesterCalendarDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 获取校历详情
      tags:
      - config
  /task/electricity:
    post:
      consumes:
      - application/json
      description: 同步电费定时推送任务
      parameters:
      - description: 任务请求内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SyncElectricityTaskRequest'
      produces:
      - application/json
      responses:
        "204":
          description: 成功，无返回内容
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 同步电费推送任务
      tags:
      - task
  /util/electricity:
    get:
      description: 根据校区、楼栋和房间号查询对应的电费余额
      parameters:
      - description: '校区名称 (例如: 云塘, 金盆岭)'
        in: query
        name: campus
        required: true
        type: string
      - description: 楼栋名称
        in: query
        name: building
        required: true
        type: string
      - description: 房间号
        in: query
        name: room
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 查询电费余额
      tags:
      - util
  /util/hello:
    get:
      description: 返回一个简单的hello world消息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Hello World测试
      tags:
      - util
  /util/profile:
    get:
      description: 使用提供的token获取用户档案信息
      parameters:
      - description: 用户Token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 获取用户个人信息
      tags:
      - util
  /util/push:
    post:
      consumes:
      - application/json
      description: 使用提供的设备Token发送推送通知
      parameters:
      - description: 推送请求内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.pushRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 发送APNS推送消息
      tags:
      - util
swagger: "2.0"